pyproject.toml
[project]
name = "ns-moon-party-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.6",
    "requests>=2.32.3",
    "solana>=0.35.0",
]
